move npcs
